Implemented review comments, create new messages file under utils
