+++ console/console-init/ui/src/utils/index.ts
@@ -5,3 +5,4 @@
 
 export * from "./utils";
 export * from "./Formatter";
+export * from "./messages";
